@@ -5,8 +5,14 @@
   "private": true,
   "scripts": {
     "start": "cross-env HTTPS=true react-app-rewired start",
+    "start:nossl": "cross-env HTTPS=false react-app-rewired start",
     "build": "react-app-rewired build",
-    "test": "react-app-rewired test",
+    "test": "react-app-rewired test --testPathIgnorePatterns 'src/visual_tests'",
+    "test:visual": "jest --runInBand --detectOpenHandles --verbose false -c src/visual_tests/jest.config.js",
+    "test:ghdiffs": "node src/visual_tests/GithubDiffUploader.js",
+    "snapshots:update": "start-server-and-test start:nossl http://localhost:3000 snapshots:capture",
+    "snapshots:capture": "jest --runInBand --detectOpenHandles --updateSnapshot -c src/visual_tests/jest.config.js",
+    "ci:nossl": "start-server-and-test start:nossl http://localhost:3000 test:visual",
     "eject": "react-scripts eject"
   },
   "author": "Lars Moastuen <lars.moastuen@cognite.com>",
@@ -18,10 +24,12 @@
   ],
   "license": "Apache-2.0",
   "dependencies": {
-    "@cognite/reveal": "../viewer/dist",
-    "@cognite/sdk": "^2.29.0",
-    "camera-controls": "^1.21.0",
+    "@cognite/sdk": "^2.32.0",
+    "camera-controls": "^1.22.1",
+    "@cognite/potree-core": "^1.1.3",
+    "@cognite/reveal": "link:../viewer/dist",
     "dat.gui": "^0.7.7",
+    "gl-matrix": "^3.3.0",
     "hold-event": "^0.0.1",
     "install": "^0.13.0",
     "react": "^16.13.1",
@@ -31,22 +39,33 @@
     "three": "^0.115.0"
   },
   "devDependencies": {
+    "@actions/core": "^1.2.4",
+    "@actions/github": "^4.0.0",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
     "@types/dat.gui": "^0.7.5",
-    "@types/jest": "^24.0.0",
-    "@types/node": "^12.0.0",
+    "@types/jest": "^26.0.3",
+    "@types/jest-environment-puppeteer": "^4.3.1",
+    "@types/jest-image-snapshot": "^3.1.0",
+    "@types/node": "^14.0.14",
+    "@types/puppeteer": "^3.0.0",
     "@types/react": "^16.9.0",
     "@types/react-dom": "^16.9.0",
     "@types/react-router-dom": "^5.1.5",
     "@types/styled-components": "^5.1.0",
     "copy-webpack-plugin": "^6.0.1",
     "cross-env": "^7.0.2",
+    "jest-environment-puppeteer": "^4.4.0",
+    "jest-image-snapshot": "^4.0.2",
+    "jest-puppeteer": "^4.4.0",
     "prettier": "^2.0.5",
+    "puppeteer": "^4.0.0",
     "react-app-rewired": "^2.1.6",
     "react-scripts": "3.4.1",
-    "typescript": "~3.7.2"
+    "start-server-and-test": "^1.11.0",
+    "ts-jest": "^26.1.1",
+    "typescript": "3.9.5"
   },
   "eslintConfig": {
     "extends": "react-app"
