@@ -15,79 +15,60 @@ describe('CogniteClient3dExtensions', () => {
     appId,
     baseUrl
   });
+  client.loginWithApiKey({ apiKey: 'dummy', project: 'unittest' });
   const clientExt = new CogniteClient3dExtensions(client);
+  const apiPath: RegExp = /\/api\/v1\/projects\/unittest\/3d\/.*/;
 
   test('getOutputs() throws error when server returns 400', async () => {
-    nock(/.*/)
-      .post(/.*/)
-      .reply(400, {});
+    const scope = nock(/.*/).get(apiPath).reply(400, {});
 
-    expect(clientExt.getOutputs({ externalId: 'externalId' })).rejects.toThrowError();
+    await expect(clientExt.getOutputs({ modelId: 1337, revisionId: 42 })).rejects.toThrowError();
+    expect(scope.isDone()).toBeTrue();
   });
 
   test('getOutputs() with empty outputs in response, returns empty list', async () => {
     // Arrange
     const response = {
-      items: [
-        {
-          model: {
-            externalId: 'externalId'
-          },
-          outputs: []
-        }
-      ]
+      items: []
     };
-    nock(/.*/)
-      .post(/.*/)
-      .reply(200, response);
+    nock(/.*/).get(apiPath).reply(200, response);
 
     // Act
-    const result = await clientExt.getOutputs({ externalId: 'externalId' });
+    const result = await clientExt.getOutputs({ modelId: 1337, revisionId: 42 });
 
     // Assert
-    expect(result).toEqual({ ...response.items[0] });
+    expect(result).toEqual({ modelId: 1337, revisionId: 42, outputs: [] });
   });
 
-  test('getOuputs() with two outputs, returns both outputs', async () => {
+  test('getOutputs() with two outputs, returns both outputs', async () => {
     // Arrange
     const response = {
       items: [
         {
-          model: {
-            id: 42
-          },
-          outputs: [
-            {
-              format: 'ept-pointcloud',
-              version: 1,
-              blobId: 1
-            },
-            {
-              format: 'ept-pointcloud',
-              version: 2,
-              blobId: 2
-            }
-          ]
+          format: 'ept-pointcloud',
+          version: 1,
+          blobId: 1
+        },
+        {
+          format: 'ept-pointcloud',
+          version: 2,
+          blobId: 2
         }
       ]
     };
-    nock(/.*/)
-      .post(/.*/)
-      .reply(200, response);
+    nock(/.*/).get(apiPath).reply(200, response);
 
     // Act
-    const result = await clientExt.getOutputs({ id: 42 });
+    const result = await clientExt.getOutputs({ modelId: 1337, revisionId: 42 });
 
     // Assert
-    expect(result).toEqual({ ...response.items[0] });
+    expect(result).toEqual({ modelId: 1337, revisionId: 42, outputs: response.items });
   });
 
   test('retrieveBinaryBlob() with binary data returns valid ArrayBuffer', async () => {
     // Arrange
     const response = '0123456789';
-    nock(/.*/)
-      .get(/.*/)
-      .reply(200, response, { 'content-type': 'binary' });
+    nock(/.*/).get(/.*/).reply(200, response, { 'content-type': 'binary' });
 
     // Act
     const result = await clientExt.getCadSectorFile(baseUrl, 'sector_5.i3d');
@@ -106,43 +87,34 @@ describe('CogniteClient3dExtensions', () => {
     const response = {
       items: [
         {
-          model: {
-            id: 42
-          },
-          outputs: [
-            {
-              format: 'unsupported-format',
-              version: 1,
-              blobId: 1
-            }
-          ]
+          format: 'unsupported-format',
+          version: 1,
+          blobId: 1
         }
       ]
     };
-    nock(/.*/)
-      .post(/.*/)
-      .reply(200, response);
+    nock(/.*/).post(apiPath).reply(200, response);
 
     // Act & Assert
     expect(
-      clientExt.getModelUrl({ modelRevision: { id: 42 }, format: File3dFormat.RevealCadModel })
+      clientExt.getModelUrl({ modelId: 1337, revisionId: 42, format: File3dFormat.RevealCadModel })
     ).rejects.toThrowError();
   });
 });
 
 describe('Model3dOutputList', () => {
   test('findMostRecentOutput() with no outputs of type, returns undefined', () => {
-    const outputs = new Model3DOutputList({ id: 42 }, [{ format: 'ept-pointcloud', version: 1, blobId: 1 }]);
+    const outputs = new Model3DOutputList(1337, 42, [{ format: 'ept-pointcloud', version: 1, blobId: 1 }]);
     expect(outputs.findMostRecentOutput('reveal')).toBeUndefined();
   });
 
   test('findMostRecentOutput() with single output of given type, returns version', () => {
-    const outputs = new Model3DOutputList({ id: 42 }, [{ format: 'ept-pointcloud', version: 1, blobId: 1 }]);
+    const outputs = new Model3DOutputList(1337, 42, [{ format: 'ept-pointcloud', version: 1, blobId: 1 }]);
     expect(outputs.findMostRecentOutput('ept-pointcloud')).not.toBeUndefined();
   });
 
   test('findMostRecentOutput() with mulitple outputs of given type, returns most recent version', () => {
-    const outputs = new Model3DOutputList({ id: 42 }, [
+    const outputs = new Model3DOutputList(1337, 42, [
       { format: 'ept-pointcloud', version: 1, blobId: 1 },
       { format: 'ept-pointcloud', version: 5, blobId: 5 },
       { format: 'ept-pointcloud', version: 3, blobId: 3 },
@@ -154,7 +126,7 @@ describe('Model3dOutputList', () => {
   });
 
   test('findMostRecentOutput() with list of supported formats, returns most recent supported', () => {
-    const outputs = new Model3DOutputList({ id: 42 }, [
+    const outputs = new Model3DOutputList(1337, 42, [
       { format: 'ept-pointcloud', version: 1, blobId: 1 },
       { format: 'ept-pointcloud', version: 5, blobId: 5 },
       { format: 'ept-pointcloud', version: 3, blobId: 3 },
