@@ -1,4 +1,7 @@
-declare module "*.jpeg" {
-   const value: any;
-   export = value;
+/*!
+ * Copyright 2020 Cognite AS
+ */
+declare module '*.jpeg' {
+  const value: any;
+  export = value;
 }
