@@ -7,5 +7,6 @@ import { SupportedModelTypes } from './types';
 export interface CogniteModelBase {
   readonly type: SupportedModelTypes;
   dispose(): void;
-  getModelBoundingBox(): THREE.Box3;
+  getModelBoundingBox(outBbox?: THREE.Box3): THREE.Box3;
+  updateTransformation(matrix: THREE.Matrix4): void;
 }
