@@ -1,5 +1,8 @@
+/*!
+ * Copyright 2020 Cognite AS
+ */
 const WorkerPlugin = require('worker-plugin');
-const WasmPackPlugin = require("@wasm-tool/wasm-pack-plugin");
+const WasmPackPlugin = require('@wasm-tool/wasm-pack-plugin');
 const path = require('path');
 const nodeExternals = require('webpack-node-externals');
 const getLogger = require('webpack-log');
@@ -22,34 +25,34 @@ function arg(env, name, defaultValue) {
   if (env[name] === undefined) {
     return defaultValue;
   }
-  if (env[name] === "true") {
+  if (env[name] === 'true') {
     return true;
   }
-  if (env[name] === "false") {
+  if (env[name] === 'false') {
     return false;
   }
   return env[name];
 }
 
 module.exports = env => {
-  const development = arg(env, "development", false);
+  const development = arg(env, 'development', false);
 
-  logger.info("Build config:");
+  logger.info('Build config:');
   logger.info(`  - development: ${development}`);
 
   const config = {
-    mode: development ? "development" : "production",
+    mode: development ? 'development' : 'production',
     entry: {
       index: './src/index.ts',
       experimental: './src/experimental.ts'
     },
-    target: "web",
+    target: 'web',
     resolve: {
       extensions: ['.tsx', '.ts', '.js'],
       symlinks: false,
       alias: {
-        '@': resolve('src'),
-      },
+        '@': resolve('src')
+      }
     },
     module: {
       rules: [
@@ -59,36 +62,30 @@ module.exports = env => {
             loader: 'ts-loader',
             options: {
               onlyCompileBundledFiles: true,
-              compilerOptions: !development ? {} : {
-                noUnusedLocals: false,
-                noUnusedParameters: false
-              }
-            },
+              compilerOptions: !development
+                ? {}
+                : {
+                    noUnusedLocals: false,
+                    noUnusedParameters: false
+                  }
+            }
           },
-          exclude: [
-            /node_modules/,
-            /src\/__tests__/,
-          ]
+          exclude: [/node_modules/, /src\/__tests__/]
         },
         {
           test: /\.(png|svg|jpg|jpeg|gif)$/,
-          use: [
-            'file-loader',
-          ],
+          use: ['file-loader']
         },
         {
           test: /\.(glsl|vert|frag)$/,
           exclude: '/node_modules/',
-          use: [
-            'raw-loader',
-            'glslify-loader'
-          ]
+          use: ['raw-loader', 'glslify-loader']
         },
         {
           test: /\.css$/,
           use: ['raw-loader']
         }
-      ],
+      ]
     },
     externals: [nodeExternals()],
     output: {
@@ -96,32 +93,29 @@ module.exports = env => {
       path: path.resolve(__dirname, 'dist'),
       sourceMapFilename: '[name].map',
       globalObject: `(typeof self !== 'undefined' ? self : this)`,
-      libraryTarget: 'umd',
+      libraryTarget: 'umd'
     },
-    devtool: development ? "inline-source-map" : "source-map",
+    devtool: development ? 'inline-source-map' : 'source-map',
     watchOptions: {
       aggregateTimeout: 1500,
-      ignored: [
-        /node_modules/,
-      ],
+      ignored: [/node_modules/]
     },
     optimization: {
-      usedExports: true,
+      usedExports: true
     },
     plugins: [
       new WasmPackPlugin({
-        crateDirectory: ".",
+        crateDirectory: '.',
         forceMode: 'production',
         watchDirectories: [
           path.resolve(__dirname, 'rust'),
           path.resolve(__dirname, '..', 'i3df', 'src'),
-          path.resolve(__dirname, '..', 'f3df', 'src'),
+          path.resolve(__dirname, '..', 'f3df', 'src')
         ]
       }),
       new WorkerPlugin()
-    ],
+    ]
   };
 
   return config;
 };
-
